refactor: --only path selection logic to core module
Moved all --only path selection and interactive filtering logic from restore.rs to a new src/core/only.rs module. Updated imports and mod.rs to reflect the new module structure, improving code organization and separation of concerns.
