+++ src/core/mod.rs
@@ -1,4 +1,5 @@
 pub mod crypto;
 pub mod indexes;
 pub mod metadata;
+pub mod only;
 pub mod permissions;
